Support more 10_10_10_2 formats as vertex buffers
Vertex input buffers only support VK_FORMAT_A2B10G10R10_UNORM_PACK32.
This cl adds all the signed and integer variants of that format.

This cl also removes StreamType and replaces it with VkFormat.

Tests: dEQP-VK.pipeline.vertex_input.*r10*
Bug: b/142661203

+++ src/Vulkan/VkFormat.h
@@ -49,8 +49,8 @@ public:
 	int blockHeight() const;
 	int bytesPerBlock() const;
 
-	int componentCount() const;
-	bool isUnsignedComponent(int component) const;
+	uint8_t componentCount() const;
+	bool isUnsignedComponent(uint8_t component) const;
 
 	int bytes() const;
 	int pitchB(int width, int border, bool target) const;
@@ -63,7 +63,7 @@ public:
 	bool has8bitTextureComponents() const;
 	bool has16bitTextureComponents() const;
 	bool has32bitIntegerTextureComponents() const;
-	bool isRGBComponent(int component) const;
+	bool isRGBComponent(uint8_t component) const;
 
 private:
 	VkFormat compatibleFormat() const;
